services/repo-worker/requirements.txt
# Async web framework
quart>=0.19.0
hypercorn>=0.16.0
uvloop>=0.19.0

# Async HTTP client (upstream proxying)
aiohttp>=3.9.0
aiohttp-retry>=2.8.0

# Async S3 storage
aiobotocore>=2.12.0
aiofiles>=23.2.0

# Authentication
PyJWT>=2.8.0
cryptography>=42.0.0

# Helm chart handling
pyyaml>=6.0.0
ruamel.yaml>=0.18.0

# Monitoring
prometheus-client>=0.20.0

# Utilities
python-dotenv>=1.0.0

# Cloud provider auth (for upstream registries)
boto3>=1.34.0
google-auth>=2.28.0
azure-identity>=1.15.0

# Type hints
typing-extensions>=4.9.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
